# openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "chdrms_server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/asset/type": {
      "get": {
        "tags": [
          "asset_type"
        ],
        "summary": "List all asset types.",
        "operationId": "list_asset_types",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetTypeDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "asset_type"
        ],
        "summary": "Create an asset type.",
        "operationId": "create_asset_type",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetType"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetTypeDto"
                }
              }
            }
          },
          "400": {
            "description": "Manufacturer by that ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/asset/type/{id}": {
      "get": {
        "tags": [
          "asset_type"
        ],
        "summary": "Get an asset type by its ID.",
        "operationId": "get_asset_type_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetTypeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Asset type by that ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "asset_type"
        ],
        "summary": "Update all fields of an asset type by its ID.",
        "operationId": "update_asset_type_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetType"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetTypeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Asset type by that ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "asset_type"
        ],
        "summary": "Delete an asset type by its ID.",
        "operationId": "delete_asset_type_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "401": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Asset type by that ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "asset_type"
        ],
        "summary": "Patch individual fields of an asset type by its ID.",
        "operationId": "patch_asset_type_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchAssetType"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetTypeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Manufacturer by that ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/group": {
      "get": {
        "tags": [
          "group"
        ],
        "summary": "List groups.",
        "operationId": "list_groups",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupInfo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "group"
        ],
        "summary": "Create group.",
        "operationId": "create_group",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupInfo"
                }
              }
            }
          }
        }
      }
    },
    "/group/{group_id}": {
      "get": {
        "tags": [
          "group"
        ],
        "summary": "Get group.",
        "operationId": "get_group_by_id",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupInfo"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssetType": {
        "type": "object",
        "required": [
          "name",
          "manufacturer"
        ],
        "properties": {
          "manufacturer": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "product_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "value": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SchemaDecimal"
              }
            ]
          }
        }
      },
      "AssetTypeDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "manufacturer"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "manufacturer": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "product_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "value": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SchemaDecimal"
              }
            ]
          }
        }
      },
      "CreateGroup": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "GroupInfo": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "PatchAssetType": {
        "type": "object",
        "properties": {
          "manufacturer": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "product_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "value": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SchemaDecimal"
              }
            ]
          }
        }
      },
...
      "SchemaDecimal": {
        "type": "number",
        "format": "float"
      },
...
    }
  },
...
}